+++ R/cbind.r
@@ -1,4 +1,0 @@
-cbind_list <- function(...){
-  cbind_list__impl(environment())  
-}
-